spawn a boss character who stare at the player
- New Character Structure
- Add Requirements in the README.md
