+++ README.md
@@ -1,7 +1,9 @@
 # Grandstand Gauntlet
 
 GameJam [Brackeys 2023](https://itch.io/jam/brackeys-9), for a week.
-The Theme is ![An end is a new beginning](https://img.itch.zone/aW1nLzExMzAyNDc0LnBuZw==/original/maBIE1.png)
+The Theme is
+
+![An end is a new beginning](https://img.itch.zone/aW1nLzExMzAyNDc0LnBuZw==/original/maBIE1.png)
 
 ## Game Ideas
 
@@ -19,6 +21,62 @@ like a spectator
 or leave to face other stronger enemy, in the next arena.
 One can decide to cut the cycle by refusing the transfer, and say:
 "Everything has an end".
+
+## Requirements
+
+### Must Have
+
+- [x] Map
+  - [x] side view
+- [ ] Enemy Boss
+  - [ ] Attack the player
+  - [ ] Big HP
+- [ ] Player
+  - [ ] Attack
+  - [x] Movement
+- [ ] Soul Shift
+  - [ ] Crowd
+  - [ ] When lethal dmg to the player -> Soul Shift
+  - [ ] Switch control to another unit
+
+### Should Have
+
+- [ ] Map
+  - [ ] Grandstand somewhere
+  - [ ] Confort Zone:
+  6/8 in the center of the screen
+  - [ ] Only move when leaving the confort zone
+  - [ ] The Gap should always be between the player and the boss.
+- [ ] Player
+  - [ ] Smooth Movement
+  - [ ] Pary
+  - [ ] Dodge / Dash
+- [ ] SFX
+  - [ ] Fight
+    - [ ] Impact / Failed Pary
+    - [ ] Successful Pary / Dodge
+- [ ] Music
+  - [ ] Fight
+
+### Could Have
+
+If Time exists in our dimension:
+
+- [ ] Music
+  - [ ] Start of Fight
+  - [ ] End of Fight
+- [ ] SFX
+  - [ ] Movement
+  - [ ] Ambiance
+    - [ ] Environment
+    - [ ] Character
+  - [ ] Crowd
+    - [ ] Acclamation of the crowd
+    - [ ] Suspense / Tension
+
+### Won't Have
+
+- [ ] Stuff we can't produce
 
 ## Name Ideas
