+++ src/main.rs
@@ -1,18 +1,17 @@
 pub mod camera;
+pub mod characters;
 pub mod constants;
 mod debug;
 mod locations;
-pub mod movement;
-mod player;
 
 use bevy::prelude::*;
 use bevy_parallax::{ParallaxCameraComponent, ParallaxPlugin};
 use bevy_rapier2d::prelude::*;
 
+use characters::CharacterPlugin;
 use constants::{CLEAR, TILE_SIZE};
 use debug::DebugPlugin;
 use locations::LocationsPlugin;
-use player::PlayerPlugin;
 
 #[rustfmt::skip]
 fn main() {
@@ -44,7 +43,7 @@ fn main() {
         ))
         .add_plugin(ParallaxPlugin)
         .add_plugin(LocationsPlugin)
-        .add_plugin(PlayerPlugin)
+        .add_plugin(CharacterPlugin)
         .add_plugin(DebugPlugin)
         .add_startup_system(spawn_camera)
         .run();
